+++ src/main/java/com/mallang/statistics/api/presentation/StatisticsController.java
@@ -1,0 +1,13 @@
+package com.mallang.statistics.api.presentation;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequiredArgsConstructor
+@RequestMapping("/statistics")
+@RestController
+public class StatisticsController {
+
+
+}
