+++ src/main/java/com/mallang/statistics/api/query/StatisticCondition.java
@@ -1,0 +1,12 @@
+package com.mallang.statistics.api.query;
+
+import com.mallang.statistics.api.query.support.PeriodType;
+import java.time.LocalDate;
+
+public record StatisticCondition(
+        PeriodType periodType,
+        LocalDate startDayInclude,
+        LocalDate lastDayInclude
+) {
+
+}
