+++ src/main/java/com/mallang/statistics/batch/exeution/JobExecutionRepository.java
@@ -1,0 +1,6 @@
+package com.mallang.statistics.batch.exeution;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface JobExecutionRepository extends JpaRepository<JobExecution, Long> {
+}
