+++ src/main/java/com/mallang/statistics/batch/exeution/Status.java
@@ -1,0 +1,7 @@
+package com.mallang.statistics.batch.exeution;
+
+public enum Status {
+    PROCESSING,
+    SUCCESS,
+    FAIL
+}
