--- src/main/java/com/mallang/statistics/job/exeution/JobExecutionRepository.java
@@ -1,6 +1,0 @@
-package com.mallang.statistics.job.exeution;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface JobExecutionRepository extends JpaRepository<JobExecution, Long> {
-}
