--- src/main/java/com/mallang/statistics/job/exeution/Status.java
@@ -1,7 +1,0 @@
-package com.mallang.statistics.job.exeution;
-
-public enum Status {
-    PROCESSING,
-    SUCCESS,
-    FAIL
-}
