--- src/main/java/com/mallang/statistics/presentation/StatisticsController.java
@@ -1,13 +1,0 @@
-package com.mallang.statistics.presentation;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RequiredArgsConstructor
-@RequestMapping("/statistics")
-@RestController
-public class StatisticsController {
-
-
-}
