--- src/main/java/com/mallang/statistics/presentation/StatisticsManageController.java
@@ -1,34 +1,0 @@
-package com.mallang.statistics.presentation;
-
-import com.mallang.auth.presentation.support.Auth;
-import com.mallang.statistics.presentation.request.StatisticConditionRequest;
-import com.mallang.statistics.query.StatisticQueryService;
-import com.mallang.statistics.query.response.PostViewStatisticResponse;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RequiredArgsConstructor
-@RequestMapping("/manage/statistics")
-@RestController
-public class StatisticsManageController {
-
-    private final StatisticQueryService statisticQueryService;
-
-    @GetMapping("/posts/{blogName}/{id}")
-    public ResponseEntity<List<PostViewStatisticResponse>> getPostViewStatistics(
-            @Auth Long memberId,
-            @PathVariable("blogName") String blogName,
-            @PathVariable("id") Long id,
-            @ModelAttribute StatisticConditionRequest request
-    ) {
-        return ResponseEntity.ok(
-                statisticQueryService.getPostStatistics(request.toDto(memberId, blogName, id))
-        );
-    }
-}
