--- src/main/java/com/mallang/statistics/query/StatisticQueryService.java
@@ -1,23 +1,0 @@
-package com.mallang.statistics.query;
-
-import com.mallang.statistics.query.dao.PostViewStatisticDao;
-import com.mallang.statistics.query.dto.PostViewStatisticQueryDto;
-import com.mallang.statistics.query.response.PostViewStatisticResponse;
-import com.mallang.statistics.query.support.StatisticConditionConverter;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
-
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-@Component
-public class StatisticQueryService {
-
-    private final PostViewStatisticDao postViewStatisticDao;
-
-    public List<PostViewStatisticResponse> getPostStatistics(PostViewStatisticQueryDto dto) {
-        StatisticCondition cond = StatisticConditionConverter.convert(dto.periodType(), dto.lastDay(), dto.count());
-        return postViewStatisticDao.find(dto.memberId(), dto.blogName(), dto.postId(), cond);
-    }
-}
