--- src/main/java/com/mallang/statistics/query/dto/PostViewStatisticQueryDto.java
@@ -1,16 +1,0 @@
-package com.mallang.statistics.query.dto;
-
-import com.mallang.statistics.query.support.PeriodType;
-import java.time.LocalDate;
-import lombok.Builder;
-
-@Builder
-public record PostViewStatisticQueryDto(
-        Long memberId,
-        String blogName,
-        Long postId,
-        PeriodType periodType,
-        LocalDate lastDay,
-        int count
-) {
-}
